arm64: add support for poll_idle()
CAUTION: This email originated from outside of the organization. Do not click links or open attachments unless you can confirm the sender and know the content is safe.

Polling in idle helps reduce the cost of remote wakeups: if the target
sets TIF_POLLING_NRFLAG (as it does while polling in idle), the scheduler
can do remote wakeups just by setting the TIF_NEED_RESCHED.

This contrasts with sending an IPI, and incurring the cost of handling
the cost of the interrupt on the receiver.

Enabling poll_idle() needs a cheap mechanism to do the actual polling
(via smp_cond_load_relaxed_timewait()) and TIF_POLLING_NRFLAG support.

arm64 has both of these. So, select ARCH_HAS_OPTIMIZED_POLL.

Signed-off-by: Ankur Arora <ankur.a.arora@oracle.com>